Refactor application to work with typescript
