+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 yarn.lock
 package-lock.json
 .env
+build
