+++ package.json
@@ -7,7 +7,7 @@
   "author": "pedrohcms",
   "license": "MIT",
   "scripts": {
-    "dev": "nodemon src/index.js"
+    "dev": "tsnd --respawn --transpile-only --ignore-watch ./node_modules src/index.ts"
   },
   "dependencies": {
     "@prisma/client": "^2.3.0",
@@ -20,6 +20,10 @@
   },
   "devDependencies": {
     "@prisma/cli": "^2.3.0",
-    "nodemon": "^2.0.4"
+    "@types/express": "^4.17.7",
+    "@types/jsonwebtoken": "^8.5.0",
+    "ts-node": "^8.10.2",
+    "ts-node-dev": "^1.0.0-pre.56",
+    "typescript": "^3.9.7"
   }
 }
