+++ src/app/utils/uniqueEmailValidator.ts
@@ -1,17 +1,11 @@
-const prisma = require("@prisma/client");
-const { PrismaClient } = require("@prisma/client");
+import { PrismaClient } from "@prisma/client";
 
 /**
  * Given a email return true if email is unique or false if already in use.
- * @function uniqueEmailValidator
- *
- * @param {String} email
- *
- * @returns {Boolean}
  *
  * @author Pedro Henrique Correa Mota da Silva
  */
-module.exports = async (email) => {
+export default async (email: string) => {
   const prisma = new PrismaClient();
 
   const found = await prisma.users.findOne({
