--- src/app/validators/idValidator.js
@@ -1,7 +1,0 @@
-const { celebrate, Joi, Segments } = require("celebrate");
-
-module.exports = celebrate({
-  [Segments.PARAMS]: Joi.object().keys({
-    id: Joi.number().required(),
-  }),
-});
