+++ src/app/validators/idValidator.ts
@@ -1,0 +1,7 @@
+import { celebrate, Joi, Segments } from "celebrate";
+
+export default celebrate({
+  [Segments.PARAMS]: Joi.object().keys({
+    id: Joi.number().required(),
+  }),
+});
