--- src/index.js
@@ -1,5 +1,0 @@
-require("dotenv").config();
-
-const app = require("./App");
-
-app.listen(3333);
