+++ src/index.ts
@@ -1,0 +1,6 @@
+import { config } from "dotenv";
+config();
+
+import express from "./App";
+
+express.listen(3333);
